# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: COMPAON API
  version: 1.0.0
  contact:
    email: compaon@dalavina.com
    name: Compaon
    url: 'https://dalavina.com/compaon'
  license:
    name: EULA
    url: 'https://en.wikipedia.org/wiki/End-user_license_agreement'
  description: API definition for COMPAON
servers:
  - url: 'http://52.74.156.150:3000/api/v1'
paths:
  '/exhibitions':
    get:
      security:
      - bearerAuth: []
      summary: Show list exhibitions
      description: Show list exhibitions
      operationId: getExhibitions
      parameters:
      - in: query
        name: page
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '1'
        description: The number of items to skip before starting to collect the result set
      - in: query
        name: limit
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '10'
        description: The numbers of items to return
      - in: query
        name: search
        allowReserved: true
        schema:
          type: string
          example: 'Triển lãm xúc tiến đầu'
        description: |
          Searching on allow columns:['name']:
          ```
          http://localhost:3000/api/v1/exhibitions?search=Triển lãm xúc tiến đầu
          ```
      - in: query
        name: sort_by
        allowReserved: true
        schema:
          type: string
          example: 'id:DESC'
        description: |
          Sort on allow columns:
           - ['id', 'DESC']
           - ['id', 'ASC']
      - in: query
        name: filter
        allowReserved: true
        schema:
          type: array
          items:
            type: string
            example: 'status:$eq:new'
        description: Filter by the field with syntax `column_name:operator:value`
      - in: query
        name: populate
        schema:
          type: array
          items:
            type: string
            example: 'booth_organization'
        description: |
          List of populatable columns
          | Populates          |
          |--------------------|
          | category           |
          | space              |
          | booth_organization |
          | booth_template     |
          | space_template     |
      - in: query
        name: withDeleted
        schema:
          type: string
          example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExhibitions'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
      tags:
        - Exhibition
    post:
      security:
      - bearerAuth: []
      tags:
        - Exhibition
      summary: Add a basic information of exhibition
      description: Add a basic information of exhibition
      operationId: addInfoExhibition
      requestBody:
        description: Create a basic information exhibition
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/Exhibition'
              - properties:
                  booth_number:
                    type: number
                    example: 100
                  space_template_id:
                    type: integer
                    writeOnly: true
                    example: 1
                  conference_template_id:
                    type: integer
                    writeOnly: true
                    example: 1
                  booth_organization_template_id:
                    type: integer
                    writeOnly: true
                    example: 1
                  category_id:
                    type: integer
                    writeOnly: true
                    example: 1
                  status:
                    default: new
              required:
              - name
              - exhibition_code
              - date_exhibition_start
              - date_exhibition_end
              - introduction
              - agenda
              - booth_number
              - category_id
              - space_template_id
              - conference_template_id
              - booth_organization_template_id
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Exhibition'
                - properties:
                    booth_number:
                      type: number
                      example: 100
                required:
                - id
                - name
                - exhibition_code
                - date_exhibition_start
                - date_exhibition_end
                - introduction
                - agenda
                - status
                - booth_number
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/exhibitions/{id}':
    get:
      security:
      - bearerAuth: []
      summary: Get exhibition by ID
      description: Return a single exhibition
      operationId: getExhibitionById
      parameters:
        - name: id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'booth_organization'
          description: |
            List of populatable columns:
            - category
            - space
            - space.locations
            - space.space_template
            - space.space_template.space_template_locations
            - booth_organization
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Exhibition'
                - properties:
                    booth_number:
                      type: number
                      example: 100
                required:
                  - id
                  - name
                  - exhibition_code
                  - date_exhibition_start
                  - date_exhibition_end
                  - introduction
                  - agenda
                  - status
                  - booth_number
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Exhibition
    patch:
      security:
      - bearerAuth: []
      summary: Update exhibition by ID
      description: Update exhibition
      operationId: updateExhibitionById
      parameters:
        - name: id
          description: ID of exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Update exhibition
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exhibition'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Exhibition'
                - properties:
                    booth_number:
                      type: number
                      example: 100
                required:
                  - id
                  - name
                  - exhibition_code
                  - date_exhibition_start
                  - date_exhibition_end
                  - introduction
                  - agenda
                  - booth_number
                  - status
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Exhibition
    delete:
      security:
      - bearerAuth: []
      tags:
        - Exhibition
      summary: Delete an existing exhibition
      description: Delete an existing exhibition by Id
      operationId: deleteExhibition
      parameters:
        - name: id
          description: ID to delete exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/exhibitions/{exhibition_id}/booths':
    post:
      security:
      - bearerAuth: []
      tags:
        - Exhibition
      summary: Add a booth
      description: Add a booth
      operationId: addBooth
      parameters:
        - name: exhibition_id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Create a booth
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booth'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/exhibitions/{exhibition_id}/booths/{booth_id}':
    put:
      security:
      - bearerAuth: []
      tags:
        - Exhibition
      summary: Update an existing booth
      description: Update an existing booth by Id
      operationId: updateBooth
      parameters:
        - name: exhibition_id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: booth_id
          description: ID to get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Update an existing booth
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booth'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not found
                    details: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    get:
      security:
      - bearerAuth: []
      summary: Get booth by ID
      description: Return a single booth
      operationId: getBoothById
      parameters:
        - name: exhibition_id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: booth_id
          description: ID to get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'booth_data.position_booth'
          description: |
            List of populable columns:
            - booth_images.booth_template_position
            - booth_projects.booth_template_position
            - booth_products.booth_template_position
            - booth_videos.booth_template_position
            - booth_template
            - location
            - location.space_template_location
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Exhibition
...
components:
  schemas:
    Exhibition:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Triển lãm xúc tiến đầu tư thương mại
        exhibition_code:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-zA-Z0-9_-]+$'
          example: TL0001
        date_exhibition_start:
          type: string
          minLength: 1
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-01-01T00:00:00.000Z'
        date_exhibition_end:
          type: string
          minLength: 1
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-12-01T00:00:00.000Z'
        introduction:
          type: string
          minLength: 1
          maxLength: 20000
          example: Giới thiệu công ty Tòa nhà C22-Officetel, số 22 Nguyễn Việt Hồng, phường An Phú, quận Ninh Kiều, TP. Cần Thơ
        agenda:
          type: string
          minLength: 1
          maxLength: 20000
          example: '07h45 - 08h20: Đón tiếp đại biểu, 08h20 - 08h40: Ca nhạc chào mừng'
        status:
          type: string
          enum: ['new','listing','finished']
        booth_organization_id:
          type: integer
          readOnly: true
          example: 1
        space_id:
          type: integer
          readOnly: true
          example: 1
        conference_id:
          type: integer
          readOnly: true
          example: 1
...
    ErrorResponseBody:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
    PaginatedExhibitions:
      type: object
      required:
        - page
        - limit
        - total
        - exhibitions
      properties:
        page:
          type: integer
          example: 0
          description: The number of items to skip before starting to collect the result set.
        limit:
          type: integer
          example: 10
          description: The number of items to return.
        total:
          type: integer
          format: int64
          example: 1
        exhibitions:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Exhibition'
            - properties:
                booth_number:
                  type: number
                  example: 100
              required:
                - id
                - name
                - exhibition_code
                - date_exhibition_start
                - date_exhibition_end
                - introduction
                - agenda
                - booth_number
                - status
...
    Booth:
      type: object
      required:
        - id
        - name
        - enterprise_id
        - created_by
        - location_id
        - booth_template_id
      properties:
        id:
          type: integer
          example: 1
          readOnly: true
        name:
          type: string
        enterprise_id:
          type: integer
          example: 1
        created_by:
          type: integer
          readOnly: true
          example: 1
        created_date:
          type: string
          readOnly: true
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-01-01T00:00:00.000Z'
        location_id:
          type: integer
          writeOnly: true
          example: 1
        booth_template_id:
          type: integer
          writeOnly: true
          example: 1
        live_streams:
          type: array
          items:
            $ref: '#/components/schemas/LiveStream'
        booth_images:
          type: array
          items:
            $ref: '#/components/schemas/BoothImage'
        booth_videos:
          type: array
          items:
            $ref: '#/components/schemas/BoothVideo'
        booth_projects:
          type: array
          items:
            $ref: '#/components/schemas/BoothProject'
        booth_products:
          type: array
          items:
            $ref: '#/components/schemas/BoothProduct'
...
    LiveStream:
      type: object
      required:
        - id
        - link
        - time
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        link:
          type: string
          minLength: 1
          example: 'www.goole.com'
        time:
          type: string
          minLength: 1
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-12-01T00:00:00.000Z'
    BoothImage:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothVideo:
      type: object
      required:
        - id
        - booth_template_position_id
        - video_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video selected from data center
          example: 1
        video_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothProject:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
        - title
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        view:
          type: integer
          readOnly: true
          example: 1
        like:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video or image selected from data center
        image_id:
          type: integer
          readOnly: true
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video or image selected from your local
          example: 'base64 string'
        title:
          type: string
          maxLength: 255
          example: 'Tên dự án'
        description:
          type: string
          maxLength: 20000
          example: 'Nội dung mô tả cho dự án'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
    BoothProduct:
      type: object
      required:
        - id
        - name
        - price
        - purchase_link
        - booth_template_position_id
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          example: 'mận Hà Nội'
        price:
          type: number
          format: float
          example: 35000
        purchase_link:
          type: string
          maxLength: 255
          example: 'https://link_mua_san_pham/ma_ha_noi.png'
        description:
          type: string
          maxLength: 500
          example: 'Mô tả sản phẩm'
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT